Made corrections based on comments, moved the tests from test_views.py to tests.py on each app Administracion, Becas, Captura, Familias and Indicadores
