+++ administracion/views.py
@@ -42,10 +42,11 @@ def admin_users_create(request):
             return redirect('administracion:users')
 
 
-@login_required
-@user_passes_test(is_administrador)
+# @login_required
+# @user_passes_test(is_administrador)
 def admin_focusmode(request):
     """ View to show the focus mode of a certain study.
+
     TODO. This is a dummy view, therefore it may
     be changed in the future.
     """
