+++ familias/views.py
@@ -7,6 +7,7 @@ from perfiles_usuario.utils import is_administrador
 @user_passes_test(is_administrador)
 def families_directory(request):
     """ View to show all the families and their members.
+
     TODO. This is a dummy view, therefore it may
     be changed in the future.
     """
@@ -18,6 +19,7 @@ def families_directory(request):
 @user_passes_test(is_administrador)
 def family_member(request):
     """ View to show the detail of a family member.
+    
     TODO. This is a dummy view, therefore it may
     be changed in the future.
     """
