+++ jp2_online/templates/administracion/focus_mode.html
@@ -2,7 +2,7 @@
 {% load staticfiles %}
 {% block dashboard_body %}
 
-    <a href="{% url 'administracion:focus' %}"> </a>
+    
     <h2 class="text-center"> Focus Mode #<h2>
 
 {% endblock %}
