+++ jp2_online/templates/becas/reinsciption_studies_left.html
@@ -2,7 +2,7 @@
 {% load staticfiles %}
 {% block dashboard_body %}
 
-    <a href="{% url 'becas:services' %}"> </a>
+    
     <h2 class="text-center"> Lista de Estudios que se deben marcar para Reinscripción <h2>
 
 {% endblock %}
