+++ jp2_online/templates/captura/economy.html
@@ -2,7 +2,7 @@
 {% load staticfiles %}
 {% block dashboard_body %}
 
-    <a href="{% url 'captura:income' %}"> </a>
+    
     <h2 class="text-center"> Ingresos y Egresos <h2>
 
 {% endblock %}
