+++ jp2_online/templates/captura/family.html
@@ -2,7 +2,7 @@
 {% load staticfiles %}
 {% block dashboard_body %}
 
-    <a href="{% url 'captura:family' %}"> </a>
+    
     <h2 class="text-center"> Inicio del Estudio: Familia <h2>
 
 {% endblock %}
