+++ jp2_online/templates/captura/housing.html
@@ -2,7 +2,7 @@
 {% load staticfiles %}
 {% block dashboard_body %}
 
-    <a href="{% url 'captura:housing' %}"> </a>
+    
     <h2 class="text-center"> Imágenes de la Vivienda <h2>
 
 {% endblock %}
