+++ jp2_online/templates/captura/sections.html
@@ -2,7 +2,7 @@
 {% load staticfiles %}
 {% block dashboard_body %}
 
-    <a href="{% url 'captura:sections' %}"> </a>
+    
     <h2 class="text-center"> Secciones del Estudio <h2>
 
 {% endblock %}
