+++ jp2_online/templates/captura/studies_list.html
@@ -2,7 +2,7 @@
 {% load staticfiles %}
 {% block dashboard_body %}
 
-    <a href="{% url 'captura:estudios' %}"> </a>
+    
     <h2 class="text-center"> Lista de estudios pendientes <h2>
 
 {% endblock %}
