+++ jp2_online/templates/familias/all_families.html
@@ -2,7 +2,7 @@
 {% load staticfiles %}
 {% block dashboard_body %}
 
-    <a href="{% url 'familias:all' %}"> </a>
+    
     <h2 class="text-center"> Directorio de Personas <h2>
 
 {% endblock %}
