+++ jp2_online/templates/familias/family_member.html
@@ -2,7 +2,7 @@
 {% load staticfiles %}
 {% block dashboard_body %}
 
-    <a href="{% url 'familias:member' %}"> </a>
+    
     <h2 class="text-center"> Detalle de Miembro de Familia #<h2>
 
 {% endblock %}
