+++ jp2_online/templates/indicadores/all_indicators.html
@@ -2,7 +2,7 @@
 {% load staticfiles %}
 {% block dashboard_body %}
 
-    <a href="{% url 'indicadores:all' %}"> </a>
+    
     <h2 class="text-center"> Lista de Indicadores <h2>
 
 {% endblock %}
