+++ jp2_online/templates/indicadores/indicator_detail.html
@@ -2,7 +2,7 @@
 {% load staticfiles %}
 {% block dashboard_body %}
 
-    <a href="{% url 'indicadores:detail' %}"> </a>
+    
     <h2 class="text-center"> Detalle del Indicador # <h2>
 
 {% endblock %}
